+++ Packages/com.thelegends.sound.manager/Editor/SoundManagerSetup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c98da75f000e5474c803ec3957e0146a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
