+++ Packages/com.thelegends.sound.manager/Editor/SoundManagerSetup/AudioAddressableSetup.cs.meta
