+++ Packages/com.thelegends.sound.manager/Runtime/AudioAddressableLoader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d4380e470388cad47821ac33036c1fce
+guid: dec9fc2e670ba6e40a3028f78e0b9d44
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
