+++ Packages/com.thelegends.sound.manager/Runtime/AudioMixerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 63caf90a6636fdc49a40320b38c2721d
+guid: 62f787ef589526549876cb5785381560
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
