+++ Packages/com.thelegends.sound.manager/Runtime/Prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52edb98822587484fae3c4784445f6a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
