+++ Packages/com.thelegends.sound.manager/Runtime/Prefab/SoundManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0fb03e83da6ca984f8637a0d6e479b9a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
