+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 974247fc44a0ac34faf1e62c2339ad6a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
