+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager/AudioAddressableLoader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f40946029e3338f4fabbe40ab0207438
+guid: d4380e470388cad47821ac33036c1fce
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
