+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager/AudioChannelType.cs
