+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager/AudioConstants.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f40946029e3338f4fabbe40ab0207438
+guid: 0fb71f43abf721a4eb449e17de14e856
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
