+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager/AudioMixerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f40946029e3338f4fabbe40ab0207438
+guid: 63caf90a6636fdc49a40320b38c2721d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
