+++ Packages/com.thelegends.sound.manager/Runtime/SoundManager/SoundManagerSettings.cs.meta
