+++ Packages/com.thelegends.sound.manager/Tests/AudioAddressableLoaderTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f40946029e3338f4fabbe40ab0207438
+guid: 1af08f5ab96565b46b852443255d2a04
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
