+++ Packages/manifest.json
@@ -1,10 +1,20 @@
 {
   "dependencies": {
+    "com.thelegends.unity.patterns": "1.0.1",
     "com.unity.addressables": "1.22.3",
     "com.unity.ide.rider": "3.0.34",
     "com.unity.ide.visualstudio": "2.0.22",
     "com.unity.test-framework": "1.1.33",
-    "com.unity.modules.audio": "1.0.0",
-    "com.unity.ugui": "1.0.0"
-  }
+    "com.unity.ugui": "1.0.0",
+    "com.unity.modules.audio": "1.0.0"
+  },
+  "scopedRegistries": [
+    {
+      "name": "TripSoft",
+      "url": "https://verdaccio-2.tripdev.io/",
+      "scopes": [
+        "com.thelegends"
+      ]
+    }
+  ]
 }
